feat: Improve authentication, ensure board and lists ✨
A few changes here : the way Trello authentication is done
has been change to take advantage of the native token
persistence of the Trello Client.

Once the authorization has been made, we now ensure that
the board exists, we create it if it is not the case.

We also fetch the lists related to the board.

The Trello model has been simplified, as well as the
AppContent component. The idea is to let the Trello
model handle the whole workflow and make AppContent
updated only when loading properties are changed.